Current Changes 1.0 : ATTENTION!: App structure need a big improvement -Api Change
Plans for continuing changes:
-(Further plans are in notebook)
-I need to finish the implementation of my API Service to display a list of Foods from the API request.
-Finish and improve the logic of the setFilter() in AllExerciseListFragment.kt.
-Solve the delays, when the user uses the sort option FragmentExerciseList.kt
-Do the logic for the other screens in this app.
-Write documentary to explain the implementation of firebass and improve the code.
-Add register and logIn Options with google.

+++ app/src/main/java/model/data/remote/api_model/openFoodFacts/Nutriments.kt
@@ -5,12 +5,14 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Nutriments(
-val energy: Int,
-val fat: Int,
+val energy: Double,
+val fat: Double,
+
 @Json(name ="saturated-fat")
-val saturatedFat: Int,
-val carbohydrates: Int,
-val sugars: Int,
-val proteins: Int,
-val salt: Int
+val saturatedFat: Double,
+
+val carbohydrates: Double,
+val sugars: Double,
+val proteins: Double,
+val salt: Double
 )
